// dist/deno/types.d.ts
export interface JWKParameters {
  /** JWK "kty" (Key Type) Parameter. */
  kty: string
  /** JWK "alg" (Algorithm) Parameter. */
  alg?: string
  /** JWK "key_ops" (Key Operations) Parameter. */
  key_ops?: string[]
  /** JWK "ext" (Extractable) Parameter. */
  ext?: boolean
  /** JWK "use" (Public Key Use) Parameter. */
  use?: string
  /** JWK "x5c" (X.509 Certificate Chain) Parameter. */
  x5c?: string[]
  /** JWK "x5t" (X.509 Certificate SHA-1 Thumbprint) Parameter. */
  x5t?: string
  /** "x5t#S256" (X.509 Certificate SHA-256 Thumbprint) Parameter. */
  'x5t#S256'?: string
  /** JWK "x5u" (X.509 URL) Parameter. */
  x5u?: string
  /** JWK "kid" (Key ID) Parameter. */
  kid?: string
}
// ...
export interface JWK extends JWKParameters {
  /**
   * - (EC) Curve
   * - (OKP) The Subtype of Key Pair
   */
  crv?: string
  /**
   * - (Private RSA) Private Exponent
   * - (Private EC) ECC Private Key
   * - (Private OKP) The Private Key
   */
  d?: string
  /** (Private RSA) First Factor CRT Exponent */
  dp?: string
  /** (Private RSA) Second Factor CRT Exponent */
  dq?: string
  /** (RSA) Exponent */
  e?: string
  /** (oct) Key Value */
  k?: string
  /** (RSA) Modulus */
  n?: string
  /** (Private RSA) Other Primes Info. This parameter is not supported. */
  oth?: Array<{
    d?: string
    r?: string
    t?: string
  }>
  /** (Private RSA) First Prime Factor */
  p?: string
  /** (Private RSA) Second Prime Factor */
  q?: string
  /** (Private RSA) First CRT Coefficient */
  qi?: string
  /**
   * - (EC) X Coordinate
   * - (OKP) The public key
   */
  x?: string
  /** (EC) Y Coordinate */
  y?: string

  [propName: string]: unknown
}